Modify animation functions and Added SaveSystem for saving scores

+++ src/image_puzzle.h
@@ -36,7 +36,7 @@ namespace Puzzle
 	    extern int move;        // number of moves taken
         extern int h1;          // number of hint 1 used
         extern int h2;          // number of hint 2 used
-        extern int totalHint;
+        extern float totalHint;
     }
 
     namespace Timer
@@ -82,7 +82,7 @@ namespace Puzzle
     bool tryMove(int dr, int dc); // dr = -1..1, dc = -1..1
 
     // puzzle result opacity
-    float getPercent(int totalHints);
+    float getPercent(float totalHints);
 }
 
 void theImagePuzzle(Image& myPuzzleImage, const Image& myImageChoosen, Texture& myPuzzleTexture);
